Move settings to app_settings
